Fixed useInstitutions and added tests for useInstitutions and MainPage
